requirements.txt
altair==5.3.0
APScheduler==3.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
clarifai==10.5.3
clarifai-grpc==10.6.1
click==8.1.7
colorama==0.4.6
contextlib2==21.6.0
emoji==2.12.1
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
grpcio==1.64.1
idna==3.7
inquirerpy==0.3.4
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-rapidjson==1.18
python-telegram-bot==13.15
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
schema==0.7.5
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.1
tqdm==4.66.4
tritonclient==2.47.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
dnspython==2.6.1 
pymongo==4.8.0